% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_planner.R
\name{filter_planner}
\alias{filter_planner}
\title{Filter Planner Data}
\usage{
filter_planner(planner, ...)
}
\arguments{
\item{planner}{Planner data returned from read_planner()}

\item{...}{dplyr::filter arguments}
}
\value{
A new Planner object with original plan name,
        original plan export date, and filtered plan data
}
\description{
Using dplyr, quickly filter planner data from read_planner()
}
\examples{
\dontrun{
## Basic Usage
     filter_planner(plan, Progress == "Not started", Priority == "Urgent")
}
}
