% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_planner.R
\name{get_planner}
\alias{get_planner}
\title{Get Planner Data}
\usage{
get_planner(xlsx)
}
\arguments{
\item{xlsx}{A Microsoft Excel spreadsheet exported from Planner}
}
\value{
A list of plan name, plan export date, and tibble of plan data
}
\description{
Read in Microsoft Planner data from .xlsx file
}
\examples{
\dontrun{
## Basic Usage
     get_planner("path/to/planner.xlsx")
}
}
