% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plannr.R
\name{print.plannr}
\alias{print.plannr}
\title{Print}
\usage{
\method{print}{plannr}(x, ...)
}
\arguments{
\item{x}{Object of class `plannr`}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
The planner name and export date.
}
\description{
Print method for `plannr` class
}
\examples{
\dontrun{
    ## Basic Usage
    planner <- read_planner("path/to/planner.xlsx")
    print(planner)
}
}
