% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_planner.R
\name{read_planner}
\alias{read_planner}
\title{Get Planner Data}
\usage{
read_planner(xlsx, plan_name = NA, plan_date = NA)
}
\arguments{
\item{xlsx}{A Microsoft Excel spreadsheet exported from Planner}

\item{plan_name}{String of preferred plan name. If `NA`, name of planner will be used.}

\item{plan_date}{String of preferred plane date. If `NA`, export date of planner will be used.}
}
\value{
A list of plan name, plan export date, and tibble of plan data
}
\description{
Read in Microsoft Planner data from .xlsx file
}
\examples{
\dontrun{
## Basic Usage
     read_planner("path/to/planner.xlsx")
}
}
