% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plannr.R
\name{summary.plannr}
\alias{summary.plannr}
\title{Summary}
\usage{
\method{summary}{plannr}(planner)
}
\arguments{
\item{planner}{Object of class `plannr`}
}
\value{
A summary of the planner.
}
\description{
Summary method for `plannr` class
}
\examples{
\dontrun{
    ## Basic Usage
    planner <- read_planner("path/to/planner.xlsx")
    summary(planner)
}
}
